sched: add 64-bit values
To support high bandwidth network interfaces, all rates (port,
subport level token bucket and traffic class rates, pipe level
token bucket and traffic class rates) and stats counters defined
in public data structures (rte_sched.h) are modified to support
64 bit counters.

Acked-by: Cristian Dumitrescu <cristian.dumitrescu@intel.com>

+++ examples/ip_pipeline/tmgr.h
@@ -40,7 +40,7 @@ struct tmgr_port *
 tmgr_port_find(const char *name);
 
 struct tmgr_port_params {
-	uint32_t rate;
+	uint64_t rate;
 	uint32_t n_subports_per_port;
 	uint32_t frame_overhead;
 	uint32_t mtu;
