+++ examples/qos_sched/cfg_file.c
@@ -59,67 +59,67 @@ cfg_load_pipe(struct rte_cfgfile *cfg, struct rte_sched_pipe_params *pipe_params
 
 		entry = rte_cfgfile_get_entry(cfg, pipe_name, "tb rate");
 		if (entry)
-			pipe_params[j].tb_rate = (uint32_t)atoi(entry);
+			pipe_params[j].tb_rate = (uint64_t)atoi(entry);
 
 		entry = rte_cfgfile_get_entry(cfg, pipe_name, "tb size");
 		if (entry)
-			pipe_params[j].tb_size = (uint32_t)atoi(entry);
+			pipe_params[j].tb_size = (uint64_t)atoi(entry);
 
 		entry = rte_cfgfile_get_entry(cfg, pipe_name, "tc period");
 		if (entry)
-			pipe_params[j].tc_period = (uint32_t)atoi(entry);
+			pipe_params[j].tc_period = (uint64_t)atoi(entry);
 
 		entry = rte_cfgfile_get_entry(cfg, pipe_name, "tc 0 rate");
 		if (entry)
-			pipe_params[j].tc_rate[0] = (uint32_t)atoi(entry);
+			pipe_params[j].tc_rate[0] = (uint64_t)atoi(entry);
 
 		entry = rte_cfgfile_get_entry(cfg, pipe_name, "tc 1 rate");
 		if (entry)
-			pipe_params[j].tc_rate[1] = (uint32_t)atoi(entry);
+			pipe_params[j].tc_rate[1] = (uint64_t)atoi(entry);
 
 		entry = rte_cfgfile_get_entry(cfg, pipe_name, "tc 2 rate");
 		if (entry)
-			pipe_params[j].tc_rate[2] = (uint32_t)atoi(entry);
+			pipe_params[j].tc_rate[2] = (uint64_t)atoi(entry);
 
 		entry = rte_cfgfile_get_entry(cfg, pipe_name, "tc 3 rate");
 		if (entry)
-			pipe_params[j].tc_rate[3] = (uint32_t)atoi(entry);
+			pipe_params[j].tc_rate[3] = (uint64_t)atoi(entry);
 
 		entry = rte_cfgfile_get_entry(cfg, pipe_name, "tc 4 rate");
 		if (entry)
-			pipe_params[j].tc_rate[4] = (uint32_t)atoi(entry);
+			pipe_params[j].tc_rate[4] = (uint64_t)atoi(entry);
 
 		entry = rte_cfgfile_get_entry(cfg, pipe_name, "tc 5 rate");
 		if (entry)
-			pipe_params[j].tc_rate[5] = (uint32_t)atoi(entry);
+			pipe_params[j].tc_rate[5] = (uint64_t)atoi(entry);
 
 		entry = rte_cfgfile_get_entry(cfg, pipe_name, "tc 6 rate");
 		if (entry)
-			pipe_params[j].tc_rate[6] = (uint32_t)atoi(entry);
+			pipe_params[j].tc_rate[6] = (uint64_t)atoi(entry);
 
 		entry = rte_cfgfile_get_entry(cfg, pipe_name, "tc 7 rate");
 		if (entry)
-			pipe_params[j].tc_rate[7] = (uint32_t)atoi(entry);
+			pipe_params[j].tc_rate[7] = (uint64_t)atoi(entry);
 
 		entry = rte_cfgfile_get_entry(cfg, pipe_name, "tc 8 rate");
 		if (entry)
-			pipe_params[j].tc_rate[8] = (uint32_t)atoi(entry);
+			pipe_params[j].tc_rate[8] = (uint64_t)atoi(entry);
 
 		entry = rte_cfgfile_get_entry(cfg, pipe_name, "tc 9 rate");
 		if (entry)
-			pipe_params[j].tc_rate[9] = (uint32_t)atoi(entry);
+			pipe_params[j].tc_rate[9] = (uint64_t)atoi(entry);
 
 		entry = rte_cfgfile_get_entry(cfg, pipe_name, "tc 10 rate");
 		if (entry)
-			pipe_params[j].tc_rate[10] = (uint32_t)atoi(entry);
+			pipe_params[j].tc_rate[10] = (uint64_t)atoi(entry);
 
 		entry = rte_cfgfile_get_entry(cfg, pipe_name, "tc 11 rate");
 		if (entry)
-			pipe_params[j].tc_rate[11] = (uint32_t)atoi(entry);
+			pipe_params[j].tc_rate[11] = (uint64_t)atoi(entry);
 
 		entry = rte_cfgfile_get_entry(cfg, pipe_name, "tc 12 rate");
 		if (entry)
-			pipe_params[j].tc_rate[12] = (uint32_t)atoi(entry);
+			pipe_params[j].tc_rate[12] = (uint64_t)atoi(entry);
 
 		entry = rte_cfgfile_get_entry(cfg, pipe_name, "tc 12 oversubscription weight");
 		if (entry)
@@ -266,67 +266,67 @@ cfg_load_subport(struct rte_cfgfile *cfg, struct rte_sched_subport_params *subpo
 
 			entry = rte_cfgfile_get_entry(cfg, sec_name, "tb rate");
 			if (entry)
-				subport_params[i].tb_rate = (uint32_t)atoi(entry);
+				subport_params[i].tb_rate = (uint64_t)atoi(entry);
 
 			entry = rte_cfgfile_get_entry(cfg, sec_name, "tb size");
 			if (entry)
-				subport_params[i].tb_size = (uint32_t)atoi(entry);
+				subport_params[i].tb_size = (uint64_t)atoi(entry);
 
 			entry = rte_cfgfile_get_entry(cfg, sec_name, "tc period");
 			if (entry)
-				subport_params[i].tc_period = (uint32_t)atoi(entry);
+				subport_params[i].tc_period = (uint64_t)atoi(entry);
 
 			entry = rte_cfgfile_get_entry(cfg, sec_name, "tc 0 rate");
 			if (entry)
-				subport_params[i].tc_rate[0] = (uint32_t)atoi(entry);
+				subport_params[i].tc_rate[0] = (uint64_t)atoi(entry);
 
 			entry = rte_cfgfile_get_entry(cfg, sec_name, "tc 1 rate");
 			if (entry)
-				subport_params[i].tc_rate[1] = (uint32_t)atoi(entry);
+				subport_params[i].tc_rate[1] = (uint64_t)atoi(entry);
 
 			entry = rte_cfgfile_get_entry(cfg, sec_name, "tc 2 rate");
 			if (entry)
-				subport_params[i].tc_rate[2] = (uint32_t)atoi(entry);
+				subport_params[i].tc_rate[2] = (uint64_t)atoi(entry);
 
 			entry = rte_cfgfile_get_entry(cfg, sec_name, "tc 3 rate");
 			if (entry)
-				subport_params[i].tc_rate[3] = (uint32_t)atoi(entry);
+				subport_params[i].tc_rate[3] = (uint64_t)atoi(entry);
 
 			entry = rte_cfgfile_get_entry(cfg, sec_name, "tc 4 rate");
 			if (entry)
-				subport_params[i].tc_rate[4] = (uint32_t)atoi(entry);
+				subport_params[i].tc_rate[4] = (uint64_t)atoi(entry);
 
 			entry = rte_cfgfile_get_entry(cfg, sec_name, "tc 5 rate");
 			if (entry)
-				subport_params[i].tc_rate[5] = (uint32_t)atoi(entry);
+				subport_params[i].tc_rate[5] = (uint64_t)atoi(entry);
 
 			entry = rte_cfgfile_get_entry(cfg, sec_name, "tc 6 rate");
 			if (entry)
-				subport_params[i].tc_rate[6] = (uint32_t)atoi(entry);
+				subport_params[i].tc_rate[6] = (uint64_t)atoi(entry);
 
 			entry = rte_cfgfile_get_entry(cfg, sec_name, "tc 7 rate");
 			if (entry)
-				subport_params[i].tc_rate[7] = (uint32_t)atoi(entry);
+				subport_params[i].tc_rate[7] = (uint64_t)atoi(entry);
 
 			entry = rte_cfgfile_get_entry(cfg, sec_name, "tc 8 rate");
 			if (entry)
-				subport_params[i].tc_rate[8] = (uint32_t)atoi(entry);
+				subport_params[i].tc_rate[8] = (uint64_t)atoi(entry);
 
 			entry = rte_cfgfile_get_entry(cfg, sec_name, "tc 9 rate");
 			if (entry)
-				subport_params[i].tc_rate[9] = (uint32_t)atoi(entry);
+				subport_params[i].tc_rate[9] = (uint64_t)atoi(entry);
 
 			entry = rte_cfgfile_get_entry(cfg, sec_name, "tc 10 rate");
 			if (entry)
-				subport_params[i].tc_rate[10] = (uint32_t)atoi(entry);
+				subport_params[i].tc_rate[10] = (uint64_t)atoi(entry);
 
 			entry = rte_cfgfile_get_entry(cfg, sec_name, "tc 11 rate");
 			if (entry)
-				subport_params[i].tc_rate[11] = (uint32_t)atoi(entry);
+				subport_params[i].tc_rate[11] = (uint64_t)atoi(entry);
 
 			entry = rte_cfgfile_get_entry(cfg, sec_name, "tc 12 rate");
 			if (entry)
-				subport_params[i].tc_rate[12] = (uint32_t)atoi(entry);
+				subport_params[i].tc_rate[12] = (uint64_t)atoi(entry);
 
 			int n_entries = rte_cfgfile_section_num_entries(cfg, sec_name);
 			struct rte_cfgfile_entry entries[n_entries];
